requirements.txt
flask
openai
graphviz
gunicorn
google-cloud-secret-manager
google-cloud-logging
